requirements.txt
paramiko
requests
rich
whois
dnspython
tqdm
